+++ base/deprecated.jl
@@ -193,6 +193,27 @@ end
 
 # BEGIN 1.5 deprecations
 
+"""
+    isimmutable(v) -> Bool
+!!! warning
+    Consider using `!ismutable(v)` instead, as `isimmutable(v)` will be replaced by `!ismutable(v)` in a future release. (Since Julia 1.5)
+Return `true` iff value `v` is immutable.  See [Mutable Composite Types](@ref)
+for a discussion of immutability. Note that this function works on values, so if you give it
+a type, it will tell you that a value of `DataType` is mutable.
+
+# Examples
+```jldoctest
+julia> isimmutable(1)
+true
+
+julia> isimmutable([1,2])
+false
+```
+"""
+isimmutable(@nospecialize(x)) = !ismutable(x)
+export isimmutable
+
+
 macro get!(h, key0, default)
     f, l = __source__.file, __source__.line
     depwarn("`@get!(dict, key, default)` at $f:$l is deprecated, use `get!(()->default, dict, key)` instead.", Symbol("@get!"))
@@ -200,5 +221,6 @@ macro get!(h, key0, default)
         get!(()->$(esc(default)), $(esc(h)), $(esc(key0)))
     end
 end
+
 
 # END 1.5 deprecations